Consolidate mock types to single file for test pkg
