--- instrumentation/database/sql/splunksql/test/driver_test.go
@@ -1,45 +1,0 @@
-// Copyright Splunk Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package test
-
-import (
-	"database/sql/driver"
-)
-
-type mockDriver struct {
-	newConnFunc func() driver.Conn
-}
-
-var (
-	_ driver.Driver        = (*mockDriver)(nil)
-	_ driver.DriverContext = (*mockDriver)(nil)
-)
-
-func newFullMockDriver() driver.Driver {
-	return &mockDriver{newConnFunc: newFullMockConn}
-}
-
-func newSimpleMockDriver() driver.Driver {
-	d := &mockDriver{newConnFunc: newSimpleMockConn}
-	return struct{ driver.Driver }{d}
-}
-
-func (d *mockDriver) Open(string) (driver.Conn, error) {
-	return d.newConnFunc(), nil
-}
-
-func (d *mockDriver) OpenConnector(string) (driver.Connector, error) {
-	return newMockConnector(d), nil
-}
