+++ instrumentation/database/sql/splunksql/test/mockdriver_test.go
@@ -19,6 +19,50 @@ import (
 	"database/sql/driver"
 )
 
+type mockDriver struct {
+	newConnFunc func() driver.Conn
+}
+
+var (
+	_ driver.Driver        = (*mockDriver)(nil)
+	_ driver.DriverContext = (*mockDriver)(nil)
+)
+
+func newFullMockDriver() driver.Driver {
+	return &mockDriver{newConnFunc: newFullMockConn}
+}
+
+func newSimpleMockDriver() driver.Driver {
+	d := &mockDriver{newConnFunc: newSimpleMockConn}
+	return struct{ driver.Driver }{d}
+}
+
+func (d *mockDriver) Open(string) (driver.Conn, error) {
+	return d.newConnFunc(), nil
+}
+
+func (d *mockDriver) OpenConnector(string) (driver.Connector, error) {
+	return newMockConnector(d), nil
+}
+
+type mockConnector struct {
+	driver *mockDriver
+}
+
+var _ driver.Connector = (*mockConnector)(nil)
+
+func newMockConnector(d *mockDriver) driver.Connector {
+	return &mockConnector{driver: d}
+}
+
+func (c *mockConnector) Connect(context.Context) (driver.Conn, error) {
+	return c.driver.newConnFunc(), nil
+}
+
+func (c *mockConnector) Driver() driver.Driver {
+	return c.driver
+}
+
 type mockConn struct{}
 
 var (
@@ -80,3 +124,64 @@ func (c *mockConn) BeginTx(context.Context, driver.TxOptions) (driver.Tx, error)
 func (c *mockConn) ResetSession(context.Context) error {
 	return nil
 }
+
+type mockRows struct{}
+
+var _ driver.Rows = (*mockRows)(nil)
+
+func newMockRows() *mockRows {
+	return &mockRows{}
+}
+
+func (r *mockRows) Columns() []string         { return nil }
+func (r *mockRows) Close() error              { return nil }
+func (r *mockRows) Next([]driver.Value) error { return nil }
+
+type mockStmt struct {
+	query string
+}
+
+var (
+	_ driver.Stmt             = (*mockStmt)(nil)
+	_ driver.StmtExecContext  = (*mockStmt)(nil)
+	_ driver.StmtQueryContext = (*mockStmt)(nil)
+)
+
+func newMockStmt(query string) *mockStmt {
+	return &mockStmt{query: query}
+}
+
+func (s *mockStmt) Close() error {
+	return nil
+}
+
+func (s *mockStmt) NumInput() int {
+	return 0
+}
+
+func (s *mockStmt) Exec([]driver.Value) (driver.Result, error) {
+	return nil, nil
+}
+
+func (s *mockStmt) ExecContext(context.Context, []driver.NamedValue) (driver.Result, error) {
+	return nil, nil
+}
+
+func (s *mockStmt) Query([]driver.Value) (driver.Rows, error) {
+	return newMockRows(), nil
+}
+
+func (s *mockStmt) QueryContext(context.Context, []driver.NamedValue) (driver.Rows, error) {
+	return newMockRows(), nil
+}
+
+type mockTx struct{}
+
+var _ driver.Tx = (*mockTx)(nil)
+
+func newMockTx() *mockTx {
+	return &mockTx{}
+}
+
+func (t *mockTx) Commit() error   { return nil }
+func (t *mockTx) Rollback() error { return nil }
